# specs/001-rag-chatbot-auth/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Physical AI & Humanoid Robotics RAG Chatbot API
  description: Backend API for authenticated RAG chatbot with conversation history, recommendations, and reading progress tracking
  version: 1.0.0
  contact:
    name: API Support
    email: support@physical-ai-book.com
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.physical-ai-book.com
    description: Production server (Hugging Face Spaces)
...
security:
  - bearerAuth: []
paths:
...
  /recommendations:
    get:
      tags:
        - Recommendations
      summary: Get personalized chapter recommendations
      description: Retrieve top 5 recommended chapters based on user background and progress
      responses:
        '200':
          description: Recommendations retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  recommendations:
                    type: array
                    items:
                      $ref: '#/components/schemas/Recommendation'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /recommendations/{recommendation_id}/dismiss:
    put:
      tags:
        - Recommendations
      summary: Dismiss recommendation
      description: Mark recommendation as dismissed (will not show again)
      parameters:
        - name: recommendation_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Recommendation dismissed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Recommendation dismissed
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Recommendation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Progress Endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminAuth:
      type: apiKey
      in: header
      name: X-Admin-Key
  responses:
    UnauthorizedError:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Recommendation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        chapter_id:
          type: string
        chapter_title:
          type: string
        score:
          type: number
          format: float
          minimum: 0
          maximum: 1
        reason:
          type: string
        created_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
        details:
          type: object
          additionalProperties: true